// src/constants/Constants.d.ts
import { DateFormat, FileSize, TypeFile } from "../types/FormTypes";
export declare const DEFAULT_ERROR_PATTERN_MESSAGE: string;
export declare const DEFAULT_ERROR_REQUIRED_MESSAGE: string;
export declare const DEFAULT_ERROR_DIGITS_ONLY_MESSAGE: string;
export declare const DEFAULT_ERROR_EMAIL_MESSAGE: string;
export declare const DEFAULT_ERROR_URL_MESSAGE: string;
export declare const DEFAULT_ERROR_FILE_TYPE_MESSAGE: string;
export declare const DEFAULT_ERROR_FILE_SIZE_MESSAGE: string;
export declare const DEFAULT_ERROR_FILE_DIMENSIONS_MESSAGE: string;
export declare const DEFAULT_ERROR_FORMAT_DATE_MESSAGE: string;
export declare const DEFAULT_ERROR_NUMBER_POSITIVE: string;
export declare const DEFAULT_ERROR_NUMBER_NEGATIVE: string;
export declare const DEFAULT_ERROR_ALPHA: string;
export declare const DEFAULT_ERROR_ALPHA_NUMERIC: string;
export declare const DEFAULT_ERROR_LOWER_CASE: string;
export declare const DEFAULT_ERROR_UPPER_CASE: string;
export declare const defaultErrorMinLengthMessage: (value: number) => string;
export declare const defaultErrorMaxLengthMessage: (value: number) => string;
export declare const defaultErrorNumberRangeMessage: (min: number, max: number) => string;
export declare const defaultErrorFileDimensionsMessage: (width: number, height: number) => string;
export declare const defaultErrorFormatDateMessage: (format: DateFormat) => string;
export declare const defaultErrorExactLength: (exactLength: number) => string;
export declare const EXPRESSION_REGULAR_ONLY_NUMBERS: RegExp;
export declare const EXPRESSION_REGULAR_DECIMALS: RegExp;
export declare const EXPRESSION_EMAIL: RegExp;
export declare const EXPRESSION_URL: RegExp;
export declare const EXPRESSION_ALPHA: RegExp;
export declare const EXPRESSION_ALPHA_NUMERIC: RegExp;
export declare const EXPRESSION_LOWER_CASE: RegExp;
export declare const EXPRESSION_UPPER_CASE: RegExp;
export declare const expressionNumberRangeValidator: (value: number, min: number, max: number) => boolean;
export declare const expressionNumberPositive: (value: number) => boolean;
export declare const expressionNumberNegative: (value: number) => boolean;
export declare const expressionFileSizeValidator: (file: File, maxSize: number | FileSize) => boolean;
export declare const expressionFileTypeValidator: (file: File, allowedTypes: TypeFile[] | string[]) => boolean;
export declare const expressionImageDimensionsValidator: (file: File, dimensions: {
    width: number;
    height: number;
}) => boolean;
export declare const expressionRequiredValidator: (value: any) => boolean;
export declare const expressionMinLengthValidator: (value: string, minLength: number) => boolean;
export declare const expressionMaxLengthValidator: (value: string, maxLength: number) => boolean;
export declare const expressionDigitsOnlyValidator: (value: string) => boolean;
export declare const expressionEmailValidator: (value: string) => boolean;
export declare const expressionUrlValidator: (value: string) => boolean;
export declare const expressionAlpha: (value: string) => boolean;
export declare const expressionAlphaNumeric: (value: string) => boolean;
export declare const expressionLowerCase: (value: string) => boolean;
export declare const expressionUpperCase: (value: string) => boolean;
export declare const expressionDateFormatValidator: (value: string, format: DateFormat) => boolean;