// src/validation/Validators.d.ts
export declare enum ValidationType {
    /** Campo Obligatorio. */
    Required = "Required",
    /** Longitud mínima permitida. */
    MinLength = "MinLength",
    /** Longitud máxima permitida. */
    MaxLength = "MaxLength",
    /** Solo dígitos permitidos. */
    DigitsOnly = "DigitsOnly",
    /** Rango de números permitido. */
    NumberRange = "NumberRange",
    /** Formato de email. */
    Email = "Email",
    /** URL */
    Url = "Url",
    /** Extensión de archivo */
    FileType = "FileType",
    /** tamaño de archivo */
    FileSize = "FileSize",
    /** Dimensiones de archivo*/
    FileDimensions = "FileDimensions",
    /** Patrón Personalizado */
    Pattern = "Pattern",
    /** Formato Fecha */
    DateFormat = "DateFormat",
    /** Números Positivos */
    NumberPositive = "NumberPositive",
    /**Números Negativos */
    NumberNegative = "NumberNegative",
    /** Letras */
    Alpha = "Alpha",
    /** Letras y Números */
    AlphaNumeric = "AlphaNumeric",
    /** Letras Minúsculas */
    LowerCase = "LowerCase",
    /** Letras Mayusculas */
    UpperCase = "UpperCase"
}